+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added settings
   - Added changelog
   - Added "Reset Mop Position" button
+- Urinating button can now be binded under any key (as long as it's one of the letters, or digit)
 
 ## Beta 0.1 (07.02.2020)